requirements.txt
paho-mqtt==1.6.1
pyusb==1.2.1
